chore: Enable auto-code-splitting
Don't have to manually split route definition and route component :^)
